Switch to Tcl scripts for Vivado generation
We now use Tcl scripts to automatically generate a "Manage IP" project along
with generating all the required IP as well. The required IP is generated
from the local Vivado version used to invoke the scripts, so IP will never
be locked for local use.

The top-level Vivado project is also generated from Tcl scripts. using the
local Vivado version.

The Make system is used to neatly invoke the Tcl scripts.

This version of HDL drop the clk_wiz_mmcm IP and uses BUFR to divide the
clock.