Discard the trailing sequence number from incoming
Depending on the state of ctrl_strip_seq_en, the trailing sequence number word
from an incoming packet is discarded before the packet is forwarded to the
outside world. If asserted, the outside world receives only the data words
(i.e., first N-1 words of the packet); if not, then the packet is forwarded
as-is. The count of words in the received packet is also altered accordingly.

Tested with RTDS using ILA on Vivado 2018.3.